No idea why vscode does that hfile : c thing but it added a few more
Boost the UI vbuffer size a bit.

Get a good correct skeleton drawing.  Much much easier with those
structs public.

Draw hovered and selected and normal colours.  Will probably change
those up (and get rid of the overall rect colour).  Need to do collapse
of bones, but not sure how I want to draw that yet.

Added a little UTHash list of bone UI info like if one is selected or
not.  I'll probably use this to flag unused bones to draw them a special
colour and stuff or mark which bone is doing collision shape editing
when I get to that.