New groglibs with new clay stuff
More skeleton editor stuff, trying to get back to the functionality of
the C# version.  It draws the shape of selected bones but they are not
in the right spot, and probably scaled badly.

Going to attempt to use just a single prim object for sphere/box/cap and
just scale it to fit but it isn't working right yet.  Also the capsule
draw looks superbroken, bits are inside out.

Boosted UI verts, I think the new stuff maybe uses a bit more

My xyz axis colours were off!  Confused me greatly.  Blue is now Z and
green is Y.

Use the clay macros for the SkellyInfoBox.  Since it isn't recursing or
anything may as well make it more readable.