Switch draw order of axis and light ray to fix light ray getting scaled by the collision shape draw
Watch for no selection when assigning material (should be greyed out?)

Draw sphere collision shapes.  Was using this to test while working on
ccnogui to get the proper right->left hand conversion.  Mesh X is now
correct.

Collision shapes draw in the right spot but only because they are using
the skin's root transform.  I'm not sure this will be good enough when
it comes time to raycast.