Increase info string buf a little as the help text is getting bigger with adjust mode
Bone adjust mode

Keybindings for width/depth/length/translation along bone, radius etc
for the 3 collision shapes.  Not much of that is working yet but the
keys are hooked up.

Draw adjusting bone in gold

sGetSingleNodeSelected() for easy selected index

Lots of work in the info string to show the current hotkeys for whatever
is going on.  Mostly taken from C# colladaconvert

Change bound shape working