Turn off clay debug mode
New raycaster gump on the right side for doing raycasts against statics
or characters.  Will display bones hit etc eventually

Skelly and Ray popout keys are now F2 and F3

RayCaster module with a ray primitive and a text area for results.
Doesn't do much yet, but will eventually allow the user to adjust a ray
and see which collision shapes it hits