Getting rid of that mesh dictionary.  Characters and statics now just keep a pointer to the meshes they use
Lots of load and save stuff changes as a result and most of it is
probably broken

Statics loading fairly well and handling multiple submeshes.  There's no
reading of the individual transforms though, I think those will live in
the nodes where character bones are for skinned stuff but not sure.