Get rid of the node rename with indexes thing, that was just for testing stuff, and it screws up the collision shape mirroring
Experimented a bit with mirroring verts across x for characters.  I
think it could work if I were to fully left hand convert all the
animation stuff.

It is commented out right now and all the anims are left in right hand
space and flipped into left via a mirrored root transform (like the
collada stuff used to be)