Experimenting with customizable colours for the last Ludum Dare.  Turned out to be a pain.
I've got a fixed table of 16 colours as a static.  It is kind of game
specific but maybe I'll think of a gui for it or a better way later.

New checkbox for using vert colors as a uint index value.  If checked it
will extract the red and use it as an index.  On the blender side just
set the attribute with the hex red color being the index value.

Ditched the ambiguous vert formats and got static meshes down to a
single vec4 uint4 struct.  Characters will need a bit more work, but it
is alot less work.  The static hlsl file just has one vertex shader now
instead of 69 different ones.

This simplified the bin extraction alot as I just alloc an array of the
eventual value the shader wants and set it directly instead of all the
squishspace squishface stuff.

Using SRV / StructuredBuffer for most rendering now.  UI is still using
layouts, but I might change it later.